+++ packages/vaex-hdf5/vaex/hdf5/export.py
@@ -135,7 +135,7 @@ def export_hdf5(dataset, path, column_names=None, byteorder="=", shuffle=False, 
         logger.debug("virtual=%r", virtual)
         logger.debug("exporting %d rows to file %s" % (N, path))
         # column_names = column_names or (dataset.get_column_names() + (list(dataset.virtual_columns.keys()) if virtual else []))
-        column_names = column_names or dataset.get_column_names(virtual=virtual, strings=True)
+        column_names = column_names or dataset.get_column_names(virtual=virtual, strings=True, alias=False)
 
         logger.debug("exporting columns(hdf5): %r" % column_names)
         sparse_groups = collections.defaultdict(list)
@@ -148,7 +148,7 @@ def export_hdf5(dataset, path, column_names=None, byteorder="=", shuffle=False, 
                 sparse_matrices[id(sparse_matrix)] = sparse_matrix
                 continue
             dtype = dataset.dtype(column_name)
-            if column_name in dataset.get_column_names(virtual=False):
+            if column_name in dataset.get_column_names(virtual=False, alias=False):
                 column = dataset.columns[column_name]
                 shape = (N,) + column.shape[1:]
             else: